API overhaul
This completely changes how to load data from the API. Rather than
calling APIHelper's methods directly, APIService is used as an
intermediary. APIService calls through to APIHelper and sends a
broadcast with the API result to update any interested
Activities/Fragments.

This has the advantage of not leaking Activities because of dangling
references. This can also be easily expanded to download guides and
images for offline browsing.