Add database to store API results.
This is very much a work in progress. The results are stored so they can
be retrieved while offline. However, there isn't a way to see what you
can view offline. Currently this is merely a layer of cache on top of
doing web api requests.

Future work includes comparing dates to ensure we are not serving stale
data, login and downloading of favorited guides for offline browsing,
updating the schema to enforce uniqueness of target + query.