Add private site login for Dozuki
These changes add private sites to the app.  If the user selects a private
site, they are now prompted to log in.

Still need to test edge cases, but right now this *seems* to work.  Still need
to add a log out option to the action bar, as well as test registration or
disable registration on private sites (not sure what we want to do about that).