Fixed logout site list error
When a use logged out of a private site and was redirected to the site list
activity, the list wasn't repopulated and the APIService was throughing an
exception.  This was because the main Site object was still set to the private
site, which the app was no longer able to retrieve sites from that endpoint.
Setting the site to NULL on logout fixed this.

Once we make the Login changes we've been talking about, this change will
*hopefully* no longer be necessary.