Perform logout API call on logout
Unfortunately this doesn't work because we fire off the API call
in a service and then perform the HTTP request in an AsyncTask. By
the time we are going to include the user's authentication token
the user has been logged out of the app so we don't have the token
anymore.