Fix missing guide text
We were getting complaints of step line text not showing up on
certain devices like the S3.  I was able to reproduce it with a 2.2 moto droid
emulator.  It was caused by setting layout_width and layout_height with @dimen
constants (-1dp and -2dp) for match_parent and fill_content.