Step edit media overhaul
Step media in the guide create process was a mess.  Literally nothing was DRY,
the main image and thumbnail inflation was just duplicating ThumbnailView, and
code formatting was not followed.  This is the start at fixing all this.

I didn't intend for this commit to get so crazy, but as I was working on the
ThumbnailView conversion in the step edit fragments, I kept finding things I
wanted to change.  I know, horrible practice, but I couldn't help myself.
Because of this, the commit is a mess with way to many changes to be useful,
but below is my attempt at listing everything that is happening in here.

* Project and File organization
   * Packages guide_create and guide_view --> guide.create and guide.view
   * Removed some unused xml files (guide_step_body something)
* Abstracted out most of the guide create step media elements to use the same
  elements as the guide stpe view page.  Drastically simplified the fragment
  and DRYs everything up.
* Code formatting
* Moved "title" edit field out of the media fragment and into the step lines
  fragment.
* Misc code simplification.