Fix dialog display delay in SiteListActivity
Previously, we were waiting for the site list response to come back before
showing the dialog.  This is silly because if you are on a slow connection, the
delay between tap and dialog was 30-60 seconds and it seemed like the app was
broken.

Now, we launch the dialog right away, THEN fire off the request for the site
list, and show a nice loading indicator while we wait.