Fix OutOfMemory errors in Guide View and Edit
Views sometimes will hold onto references to bitmaps which causes a slow memory
leak.  This was causing the app to crash after paging through 20 or so steps
(depending on the device).  Releasing the bitmaps and setting the imageviews to
null frees the bitmap from native memory and allows the GC to do its thing.