Robustify login prompting
Now if the user's session expire, we prompt them to log in and perform
the API call again once they are finished logging in just like we do if
the user isn't logged in to begin with. I had to change the way we
trigger the dialog to be opened because APIService doesn't have access
to any Activities. Now we post an event to the Bus that Activities
listen for and open the dialog when it comes along.