Rework of Dozuki site list (Fixes #58)
Now SiteListActivity does the API call and sets the list on the Dialog
whenever it is needed. This has the advantage of only ever making one
API call that is initiated as soon as the app opens.

This also rejiggers site list parsing so errors percolate up to
APIService so an appropriate dialog can be displayed.