Patch/Hack to prevent segfault from OkHttp
https://github.com/square/okhttp/issues/184

Because OkHttp uses the shared SSLContext, which also happens to be used by the
Google Analytics library, it would silently and abruptly cause an segfault
breaking everything.  The above issue mentions a workaround creating a new
sslcontext and setting the OkHttpClient to use that one instead of the default
sslcontext fixes it.