Disable multiple pending guide publish API calls
Previously, the user could fire off multiple guide publish API calls by
clicking the publish button multiple times. This would cause edit
conflicts for the subsequent requests because they didn't send along the
correct revisionid.

Now the button doesn't work if we are currently changing the privacy of
a guide. Additionally, the text is changed from (Publish|Unpublish) to
(Publishing|Unpublishing) to indicate that the user's action is being
processed. A spinner would be nice but not right now.