Remove artificial character limit on step line
This was causing problems when a step line was longer than 350 characters due
to wiki text links, the edit text field would truncate the text to 350 which is
a bad call.  We shouldn't be doing any limiting client side, instead we should
just handle the validation error the api comes back with.