Do not finish activity on 403 responses
Anytime you got an insufficient priviledge error (403) from doing something
like changing the privacy of a guide (public to private), the activity would
finish after clicking OK on the dialog, which is unexpected for the user.

This commit also fixes some inconsistent UI issues as a result of failing
public/private toggles so after the request fails, the UI is updated to its
original state.