Site list: Various search fixes
We weren't resetting the visibility of some views after we receive the
site list. They, the search view in particular, were stuck as hidden and
would never be displayed.

Make the SearchView always expanded since we have the full width of the
dialog allocated for the view.

Fix crash on orientation changes with search text present. Calling
handleIntent from onCreate was a bad idea because it called some UI
changing code in the dialog before the dialog was ready for it.

Also removed the subscribe since it doesn't receive Otto events.