Fix topic hierarchy on tablets
The topic hierarchy pane wasn't behaving properly - it was going back
when it was supposed to and wouldn't close if the right side of the
screen was tapped. Turns out the back stack listener was more important
than I thought. Even though we don't do as much as we used to with
mBackStackSize, we still need to listen for changes and show/hide the
topic list as necessary.