Fix Logout events getting stuck in DeadEvents
Now we don't post Logout APIEvents to the Bus because a) we don't care
about the response which means that b) we don't listen for it.
Previously they were being posted and would get stuck in APIService's
list of DeadEvents because the Activity that it is associated with is
dead since they just logged out.