Fix #150 Gallery closing on orientation changes
The base class's onLogin method was never called in GalleryActivity
because GalleryActivity defined a method of the same signature that
overrides the base class's. That method was being called twice -- once
for the @Subscribe and once for the @Subscribe in the object listener in
BaseActivity.

Overriding BaseActivity's onLogin method and calling super fixes it.