Super rough initial implementation of protobufs
All endpoints are now broken but it's printing out an image field from a
protobuf! It'll be pretty easy to refactor the code if all of the
endpoints use protobufs. It would be pretty lame to use our API half on
JSON and half on protobufs so this shouldn't be a problem.