Display error Toast if launching the camera fails
Some devices require physical external storage to request the external
storage directory. When it doesn't exist, native apps such as the camera
decide to stop the user from taking pictures because there isn't a place
to store it. We weren't displaying an error message in this case so it
was really awkward to the user.