Handle changed password reauthentication
This is almost gracefully handled -- the user is prompted to
reauthenticate if the automatic reauthentication didn't work, the
account is left with an invalidated auth token, when the new auth token
is created, all account info is updated with any changes.

The only problem is if the user signs into a different account than the
one we already have. It needs to delete the old one and create a new one
rather than updating the old one.