Update credentials when retrying API calls
Because we just got a new user, we need to update the credentials for
the pending API call so it has an updated auth token. Otherwise an extra
login call is performed since the auth token is the old, invalid one.
However, his time the call succeeds because the password is up to date.
Updating the auth token is easier than relying on that series of events.