Rework image downloading to get accurate progress
This addresses some issues present in the previous image downloading
implementation of using a single Set to store all of the images needing
downloading.

To display realtime progress to the user, we must find all images that
need to be downloaded to attain a count. Then we must download images
one by one at a guide-level to make progress on a guide because it
doesn't make sense to download images from guides in an arbitrary order.
It makes much more sense to download all of a guide's images then move
on to the next guide. The total image count will be used for the
notification.