Initial implementation of viewing offline guides
It involved passing an offline guide flag through a few levels of
adapters/fragments/views to get it to the right place. Also, switched
some of the guide media fragments to use the newInstance() idiom.

The thumbnail code is pretty tricky because it works for viewing and
editing. It's very difficult to determine what tags certain views have
and if the value has the size or not. It's entirely possible that I
broke something with it. We'll have to do extensive testing to find all
the bugs.