Fix strange crash when viewing video
We were getting a "Class not found when unmarshalling:
android.support.v4.app.FragmentManager State" exception when viewing a guide
video in portrait mode and then canceling out of the video then going back a
page.

Setting `setSaveEnabled` to false on the ViewPager fixes this issue, for some
reason.  Source:
http://stackoverflow.com/questions/11381470/classnotfoundexception-when-unmarshalling-android-support-v4-view-viewpagersav