Add sync restart functionality
The issue was that favoriting multiple guides in a row resulted in sync
progress that wasn't correct. If the first guide takes an hour to sync,
any guide favorited before it finishes won't show up in the list or will
be taken into account for the progress because the sync adapter doesn't
know that there are more guides to sync. Once it is done, the sync
requests go through and start it so they will be synced eventually but
the progress isn't correct which is very misleading.