SSO login: Fix webview
I'm not sure what exactly what caused the issue, but the WebView used to
display OpenID/SSO login wasn't displaying any content.

I think it was ultimately a combination of the following:

 - Upgrading to 4.4 which uses Chrome for the default WebView client
 - Loading pages with insecure content
 - Layout width
 - Overriding the WebChromeClient
 - Not specifying a wide viewport

At any rate, this code is more consistent with the other WebView we use
for displaying Answers and no longer exhibits any of the aforementioned
problems.