IntentFilter: Fix guide view orientation crash
1. Navigate to a guide through the IntentFilterActivity (scan a QR code
   or view a guide/category/search URL).
1. Change orientation.
1. Crash

I couldn't ever fully track down the issue, but doing the above resulted
in a DeadEvent _before_ `onGuide` was eventually called. The DeadEvent
was retried when changing orientation which happens before onCreate.
This resulted in a crash because the ViewPager wasn't initialized yet.
Removing the call to `finish()` fixed the DeadEvent issue and
subsequently the orientation issue.