Barcode scanner: Fix failing scan result parsing
1. Go to the Media Manager.
1. Add an image from the Gallery.
1. A "Failed to parse result" toast is displayed but everything works as
   expected.

This is caused by the ZXing library parsing the Activity result but
returning a null `IntentResult` because the request codes didn't match.
The app didn't crash because the parsing code is all wrapped in a
try/catch to swallow all errors because of the tricky reflection. This
fixes the bug which removes the awkward Toast.