Disable sync button if no connection
This moves the sync command text update code into the handler that runs
(currently) every minute to update the last sync time. This allows us to
update the sync button's color and text to reflect changes in internet
connection.

Also, decrease the refresh interval to half a minute to make updates
happen more frequently. This can probably be decreased even more but
usually it doesn't result in any changes.