Offline guides: Swallow and report errors
The current code didn't handle certain errors from API calls gracefully.
It was possible to get a NullPointerException from the favorites
retriving code if the response is garbage (e.g. "<DOCTYPE ...") and the
status code is a 403 or 404. This resulted in a force close with the
notification left in the indeterminate state.

This swallows all Exceptions so we can clean up after ourselves. We of
course report the exception so we are aware of it.