Guide view: Fix error state
We try to display the offline version of a guide if there is an error
when retrieving content from the API or if the user doesn't have
internet. If the offline version doesn't exist, we display the guide
from the previous APIResult just like what what have happened normally.
Except that there isn't a guide if there was an error in the API e.g.
5xx or some sort of client error.

This fixes the display code so it goes through the same checks as the
original API result handling code.