Prepare project for Android Studio 1.0 Upgrade
Android studio is finally out of Beta - which means upgrade time!  Included
with the 1.0 release, the android build team released gradle-android-plugin
1.0.0 (http://tools.android.com/tech-docs/new-build-system/migrating-to-1-0-0)
  as well as a new version of the android support library (v21).

Luckily, there were only a couple renamed properties in the new version, so the
changes are minor.

Also important to note - you need to have `gradle` v2.2.1 installed in order to
build the project.