IntentFilterActivity: Force new task
In order to make the back button behave properly when coming from search
results or other intents, we must force `IntentFilterActivity` to have a
different `taskAffinity` so it doesn't launch the activities into the
same stack as the main application. This results in an empty back stack
for all Activities routed through `IntentFilterActivity` which is
exactly what we want.