Major overhaul of app to meet new standards
Not completely finished, but it's gotten a bit out of control with these
changes.  I should have been committing all along, breaking things up in
chunks, but :shrug: I didn't... A number of these changes were made as a result
of unexpected problems after upgrading a package or using a new AppCompat
Library.

Here's a few of the things that have been changed in this commit:
- Use the new RecyclerView for guide lists on both the view and edit sides
- Use CoordinatorLayouts and the design libraries for managing layouts
- Add some FABs
- Fix the bug where the media manager would let you add images from newer
  phones
- Upgraded to OkHttp3, removed dependency on HttpRequest
- Migrated our offline cache from using a custom store to using OkHttp3 and
  Picassos  cache so that all images will be cached, not just guides.  It also
  simplifies the implementation.  A couple of bugs still, but the majority of
  it works.
- Migrated all icons from png drawables to new Vector drawables
- Use as much of the new material design as possible.  Still lots more to do.