Fix a CommentsActivity crash when visiting a guide with lots of comments
If a guide had 100+ comments (for example, iPhone 5 battery repair), the activity would crash
because we were passing along an ArrayList<Comment> object prepopulated
with the comment information.

This was causing a TransactionToLarge exception, which android 7+ is now crashing on.

There wasn't a good reason for us to not just re-fetch the guide api.  Varnish is in front of it
and it doesn't happen all that often.